+++ axum/src/api/core/learning.rs
@@ -1,6 +1,7 @@
 use crate::api::error::APIError;
 use crate::auth::jwt::Claims;
-use crate::models::cards_decks::{CardProgress, CardProgressWithFields, ReviewPayload};
+use crate::db::crud::learning;
+use crate::models::{CardProgressWithFields, ReviewPayload, UpdateCardProgress};
 use crate::schema::AppState;
 use crate::tools::sm2::SM2Calculator;
 use axum::extract::{Json, Path, State};
@@ -11,26 +12,9 @@ pub async fn fetch_due_cards(
     State(state): State<AppState>,
     claims: Claims,
 ) -> Result<Json<Vec<CardProgressWithFields>>, APIError> {
-    let progress_list = sqlx::query_as!(
-        CardProgressWithFields,
-        r#"
-        SELECT
-            cp.*,
-            c.front,
-            c.back,
-            c.media_url
-        FROM card_progress cp
-        JOIN cards c ON c.id = cp.card_id
-        WHERE cp.user_id = $1
-            AND (cp.due_date <= CURRENT_TIMESTAMP OR cp.review_count = 0)
-        ORDER BY cp.due_date ASC
-        "#,
-        claims.sub,
-    )
-    .fetch_all(&state.db)
-    .await?;
+    let due = learning::learning::fetch_due(&state.db, &claims.sub).await?;
 
-    Ok(Json(progress_list))
+    Ok(Json(due))
 }
 
 pub async fn update_card_progress(
@@ -41,21 +25,8 @@ pub async fn update_card_progress(
 ) -> Result<StatusCode, APIError> {
     let calculator = SM2Calculator::default();
 
-    // Get current progress without transaction
-    let current_progress = sqlx::query_as!(
-        CardProgress,
-        r#"
-        SELECT cp.* FROM card_progress cp
-        WHERE cp.user_id = $1 AND cp.id = $2
-        "#,
-        claims.sub,
-        card_id
-    )
-    .fetch_one(&state.db)
-    .await
-    .map_err(|_| APIError::NotFound("Card progress not found".into()))?;
+    let current_progress = learning::learning::find_by_id(&state.db, &card_id, &claims.sub).await?;
 
-    // Calculate new values
     let (new_ease, new_interval, new_review_count) = calculator.calculate_next_review(
         payload.quality,
         current_progress.ease_factor,
@@ -63,30 +34,14 @@ pub async fn update_card_progress(
         current_progress.review_count,
     );
 
-    let now = OffsetDateTime::now_utc();
-    let new_due_date = now + Duration::days(new_interval.into());
-
-    // Simple update without returning data
-    sqlx::query!(
-        r#"
-        UPDATE card_progress SET
-            review_count = $1,
-            ease_factor = $2,
-            interval = $3,
-            last_reviewed = $4,
-            due_date = $5
-        WHERE user_id = $6 AND id = $7
-        "#,
-        new_review_count,
-        new_ease,
-        new_interval,
-        now,
-        new_due_date,
-        claims.sub,
-        card_id
-    )
-    .execute(&state.db)
-    .await?;
+    let update = UpdateCardProgress {
+        review_count: new_review_count,
+        ease_factor: new_ease,
+        interval: new_interval,
+        last_reviewed: OffsetDateTime::now_utc(),
+        due_date: OffsetDateTime::now_utc() + Duration::days(new_interval.into()),
+    };
+    learning::learning::update(&state.db, &card_id, &claims.sub, update).await?;
 
     Ok(StatusCode::OK)
 }
@@ -96,49 +51,7 @@ pub async fn reset_deck_progress(
     claims: Claims,
     Path(deck_id): Path<String>,
 ) -> Result<StatusCode, APIError> {
-    let mut tx = state.db.begin().await?;
-
-    let deck_exists = sqlx::query!(
-        r#"
-        SELECT 1 as "exists!: bool" FROM decks
-        WHERE id = $1 AND (
-            created_by = $2
-            OR assignee = $2
-            OR visibility = 'public'
-        )
-        "#,
-        deck_id,
-        claims.sub
-    )
-    .fetch_optional(&mut *tx)
-    .await?
-    .is_some();
-
-    if !deck_exists {
-        return Err(APIError::NotFound("Deck not found".into()));
-    }
-
-    sqlx::query!(
-        r#"
-        UPDATE card_progress cp
-        SET
-            review_count = 0,
-            ease_factor = 2.5,
-            interval = 1,
-            last_reviewed = NULL,
-            due_date = CURRENT_TIMESTAMP
-        FROM cards c
-        WHERE cp.card_id = c.id
-        AND c.deck_id = $1
-        AND cp.user_id = $2
-        "#,
-        deck_id,
-        claims.sub
-    )
-    .execute(&mut *tx)
-    .await?;
-
-    tx.commit().await?;
+    learning::learning::reset(&state.db, &deck_id, &claims.sub).await?;
 
     Ok(StatusCode::OK)
 }
