+++ axum/src/api/core/mod.rs
@@ -1,11 +1,9 @@
 pub mod decks;
 pub mod learning;
 pub mod lesson;
-pub mod notes;
 pub mod task;
 
 pub use decks::*;
 pub use learning::*;
 pub use lesson::*;
-pub use notes::*;
 pub use task::*;
