+++ axum/src/api/core/notes.rs
@@ -1,3 +1,5 @@
+//not used
+
 use crate::api::error::APIError;
 use crate::auth::jwt::Claims;
 use crate::models::lessons::{StudentNote, StudentNoteUpdate};
