+++ axum/src/api/routes/file_routes.rs
@@ -1,17 +1,13 @@
 use crate::api::files::file;
 use crate::schema::AppState;
-use axum::routing::{delete, get, post};
+use axum::routing::get;
 use axum::Router;
 
 pub fn file_routes() -> Router<AppState> {
-    Router::new()
-        .route("/", get(file::list_files))
-        .route(
-            "/{file_id}",
-            get(file::fetch_file)
-                .patch(file::update_file)
-                .delete(file::delete_file),
-        )
-        .route("/folder", post(file::create_folder))
-        .route("/folder/{folder_id}", delete(file::delete_folder))
+    Router::new().route("/", get(file::list_files)).route(
+        "/{file_id}",
+        get(file::fetch_file)
+            .patch(file::update_file)
+            .delete(file::delete_file),
+    )
 }
