+++ axum/src/api/routes/mod.rs
@@ -1,10 +1,9 @@
 pub mod auth_routes;
+pub mod deck_routes;
+pub mod file_routes;
 pub mod lesson_routes;
-pub mod notes_routes;
+pub mod profile_routes;
+pub mod s3_routes;
+pub mod student_routes;
 pub mod task_routes;
 pub mod user_routes;
-pub mod student_routes;
-pub mod profile_routes;
-pub mod deck_routes;
-pub mod s3_routes;
-pub mod file_routes;
