+++ axum/src/api/routes/notes_routes.rs
@@ -1,3 +1,5 @@
+//not used
+
 use crate::api::core;
 use crate::schema::AppState;
 use axum::routing::get;
