+++ axum/src/db/crud/learning/mod.rs
@@ -1,3 +1,4 @@
 pub mod card;
 pub mod deck;
+pub mod learning;
 pub mod subscribe;
